Start of OO refactoring
